+++ packages/client/src/components/code-workspace/CodeWorkspace.tsx
@@ -1,13 +1,14 @@
 import { observer } from "mobx-react-lite";
+import { AppFileEditor } from "@/components/app-workspace/app-file-editor";
+import { AppFileExplorer } from "@/components/app-workspace/app-file-explorer";
+import { useWorkspace } from "@/hooks";
 import { SettingsNavPanel } from "../../components/blocks-workspace/panels";
 import {
-	FileEditorPanel,
-	FileExplorerPanel,
 	SettingsPanel,
 	TerminalPanel,
-	Workspace,
+	WorkspaceManager,
 } from "../../components/workspace";
-import type { WorkspaceOptions, WorkspaceStore } from "../../stores";
+import type { WorkspaceOptions } from "../../stores";
 import { MCPJsonEditor } from "../workspace/panels/MCPJsonEditor";
 import { CodeWorkspaceActions } from "./CodeWorkspaceActions";
 import { RendererPanel } from "./panels";
@@ -32,7 +33,7 @@ const DEFAULT_OPTIONS: WorkspaceOptions = {
 						id: "file-explorer",
 						type: "tab",
 						name: "Files",
-						component: "file-explorer",
+						component: "app-file-explorer",
 						enableClose: false,
 						config: {},
 					},
@@ -85,29 +86,30 @@ const DEFAULT_OPTIONS: WorkspaceOptions = {
 	},
 };
 
-interface CodeWorkspaceProps {
-	/** Workspace to render */
-	workspace: WorkspaceStore;
-}
-
 /**
  * Render the code workspace
  */
-export const CodeWorkspace = observer((props: CodeWorkspaceProps) => {
-	const { workspace } = props;
-	const FACTORY: React.ComponentProps<typeof Workspace>["factory"] = (
+export const CodeWorkspace: React.FC = observer(() => {
+	const FACTORY: React.ComponentProps<typeof WorkspaceManager>["factory"] = (
 		node,
 		layout,
 	) => {
+		const { workspace } = useWorkspace();
 		const component = node.getComponent();
 		const config = node.getConfig();
 
-		if (component === "file-explorer") {
-			return <FileExplorerPanel title={"Files"} layout={layout} />;
-		} else if (component === "file-editor") {
-			return <FileEditorPanel path={config.path} />;
+		if (component === "app-file-explorer") {
+			return (
+				<AppFileExplorer
+					node={node}
+					layout={layout}
+					app={workspace.appId}
+				/>
+			);
+		} else if (component === "app-file-editor") {
+			return <AppFileEditor node={node} app={workspace.appId} />;
 		} else if (component === "mcpJsonEditor") {
-			return <MCPJsonEditor dataMap={config.data}/>;
+			return <MCPJsonEditor dataMap={config.data} />;
 		} else if (component === "renderer") {
 			return <RendererPanel />;
 		} else if (component === "settingsPanel") {
@@ -120,11 +122,11 @@ export const CodeWorkspace = observer((props: CodeWorkspaceProps) => {
 
 		return <>{component}</>;
 	};
+
 	return (
-		<Workspace
+		<WorkspaceManager
 			navbarActions={<CodeWorkspaceActions />}
 			options={DEFAULT_OPTIONS}
-			workspace={workspace}
 			factory={FACTORY}
 		/>
 	);
