--- packages/client/src/components/common/File/FileEditor.tsx
@@ -1,534 +1,0 @@
-import type { OnMount } from "@monaco-editor/react";
-import prettier from "prettier";
-import parserBabel from "prettier/parser-babel";
-import parserHtml from "prettier/parser-html";
-import parserCss from "prettier/parser-postcss";
-import {
-	forwardRef,
-	lazy,
-	Suspense,
-	useEffect,
-	useImperativeHandle,
-	useMemo,
-	useRef,
-	useState,
-} from "react";
-import { runPixel } from "@semoss/sdk/react";
-import { LoadingScreen, styled, useNotification } from "@semoss/ui";
-import { languageConfigs } from "./FileEditorLanguageConfig";
-
-const Editor = lazy(() => import("@monaco-editor/react"));
-
-const IS_PRODUCTION = !import.meta.env.DEV;
-
-const StyledContainer = styled("div")(({ theme }) => ({
-	display: "flex",
-	flexDirection: "column",
-	alignItems: "center",
-	height: "100%",
-	width: "100%",
-	background: theme.palette.background.paper,
-	overflow: "hidden",
-}));
-
-interface FileEditorProps {
-	/** Type of file opened */
-	type: "app" | "insight" | "engine";
-
-	/** Space where the file is located */
-	space: string;
-
-	/** Path to the file file */
-	path: string;
-
-	/** insight id */
-	insightId?: string | null;
-	/** Refresh content trigger */
-	refreshContent?: number;
-
-	/**
-	 * Optional Model Engine to use
-	 */
-	agentModelEngine?: string;
-
-	/**
-	 *
-	 * @param isModified
-	 * @returns
-	 */
-	onChange: (content: string, isModified: boolean) => void;
-}
-
-export interface FileEditorRefDef {
-	/**
-	 *
-	 * @returns Save the file
-	 */
-	saveFile: () => Promise<void>;
-
-	/**
-	 *
-	 * @returns Format the file
-	 */
-	formatFile: () => Promise<void>;
-}
-
-export const FileEditor = forwardRef<FileEditorRefDef, FileEditorProps>(
-	function FileEditor(props, ref) {
-		const {
-			type = "app",
-			space = "",
-			path = "",
-			insightId = null,
-			agentModelEngine = "",
-			refreshContent = 1,
-			onChange = () => null,
-		} = props;
-
-		const notification = useNotification();
-		const [content, setContent] = useState("");
-		const [initialContent, setInitialContent] = useState("");
-		const [isLoading, setIsLoading] = useState(false);
-		const [LLMActionAdded, setLLMActionAdded] = useState(false);
-		// tracks filetype to address bug when prompting LLM - re-address if/when filetype added to LLM pixel
-		const wordWrapRef = useRef(false);
-		const editorRef = useRef(null);
-		const [isModified, setIsModified] = useState(false);
-
-		// update whenever the content changes
-		useImperativeHandle(ref, () => {
-			return {
-				saveFile: saveFile,
-				formatFile: formatFile,
-			};
-		}, [content]);
-
-		/**
-		 * Set the initial file
-		 */
-		useEffect(() => {
-			// load when the type space or path change
-			//if (path && !path.endsWith("/")) {
-			loadFile();
-			//}
-		}, [type, space, path, refreshContent]);
-
-		/**
-		 * Trigger the on change function
-		 */
-		useEffect(() => {
-			onChange(content, isModified);
-		}, [isModified]);
-
-		const fileLanguage = useMemo<
-			| "typescript"
-			| "javascript"
-			| "html"
-			| "css"
-			| "scss"
-			| "python"
-			| "java"
-			| "mdx"
-			| "markdown"
-			| "txt"
-			| ""
-		>(() => {
-			const ext = path.split(".").pop();
-
-			if (ext === "ts" || ext === "tsx") {
-				return "typescript";
-			}
-			if (ext === "js" || ext === "jsx") {
-				return "javascript";
-			}
-			if (ext === "html") {
-				return "html";
-			}
-			if (ext === "css") {
-				return "css";
-			}
-			if (ext === "scss") {
-				return "scss";
-			}
-			if (ext === "py" || ext === "python") {
-				return "python";
-			}
-			if (ext === "java") {
-				return "java";
-			}
-			if (ext === "mdx") {
-				return "mdx";
-			}
-			if (ext === "md") {
-				return "markdown";
-			}
-			if (ext === "txt") {
-				return "txt";
-			}
-
-			return "";
-		}, [path]);
-
-		/**
-		 * Load the File
-		 * @param type - type of the file
-		 * @param id - ID of the file
-		 * @param path - path to the file
-		 */
-		const loadFile = async () => {
-			try {
-				setIsLoading(true);
-				let pixel = "";
-				if (type === "app") {
-					pixel = `GetAsset(filePath=["${path}"], space=["${space}"]);`;
-				} else if (type === "engine") {
-					pixel = `GetEngineAssets(filePath=["${path}"], engine=["${space}"]);`;
-				} else if (type === "insight") {
-					throw Error("TODO");
-					// TODO: add insight
-					// pixel = `GetAsset(filePath=["${path}"], space=["${id}"]);`;
-				}
-
-				if (!pixel) {
-					throw new Error("Error missing pixel to get file");
-				}
-
-				const response = await runPixel<[string]>(pixel, insightId);
-
-				// set the content
-				const content = response.pixelReturn[0].output;
-				setContent(content);
-				setInitialContent(content);
-			} catch (e) {
-				notification.add({
-					color: "error",
-					message: e.message,
-				});
-
-				console.error(e);
-			} finally {
-				setIsLoading(false);
-			}
-		};
-
-		/**
-		 * @name formatFile
-		 * Use custom parsers to format file
-		 * TODO: Save custom configs?
-		 */
-		const formatFile = async () => {
-			try {
-				let formatted = content;
-
-				if (fileLanguage === "python") {
-					//TODO:: Implement
-				} else {
-					const prettierConfig = {};
-
-					// parsers for other languages are needed
-					if (fileLanguage === "html") {
-						prettierConfig["parser"] = "html";
-						prettierConfig["plugins"] = [parserHtml];
-					} else if (
-						fileLanguage === "javascript" ||
-						fileLanguage === "typescript"
-					) {
-						prettierConfig["parser"] = "babel";
-						prettierConfig["plugins"] = [parserBabel];
-					} else if (
-						fileLanguage === "css" ||
-						fileLanguage === "scss"
-					) {
-						prettierConfig["parser"] = "css";
-						prettierConfig["plugins"] = [parserCss];
-					}
-
-					// If we have a configuration for the selected language
-					if (Object.keys(prettierConfig).length) {
-						formatted = prettier.format(content, prettierConfig);
-					}
-
-					setContent(formatted);
-				}
-
-				notification.add({
-					color: "success",
-					message: "Success formatting file",
-				});
-			} catch (e) {
-				notification.add({
-					color: "error",
-					message: e.message,
-				});
-
-				console.error(e);
-			} finally {
-				setIsLoading(false);
-			}
-		};
-
-		/**
-		 * Save the File
-		 */
-		const saveFile = async () => {
-			const content = editorRef.current.getValue();
-			try {
-				// setIsLoading(true);
-
-				let pixel = "";
-				if (type === "app") {
-					pixel = `
-                SaveAsset(fileName=["${path}"], content=["<encode>${content}</encode>"], space=["${space}"]); 
-                CommitAsset(filePath=["${path}"], comment=["Save from editor"], space=["${space}"])
-            `;
-				} else if (type === "engine") {
-					console.log(path, "path");
-					pixel = `
-                SaveEngineAssets(filePath=["${path}"], content=["<encode>${content}</encode>"], engine=["${space}"]); 
-                CommitAsset(filePath=["${path}"], comment=["Save from editor"], engine=["${space}"])`;
-				} else if (type === "insight") {
-					throw Error("TODO");
-					// TODO: add insight
-					//     pixel = `
-					//     SaveAsset(fileName=["${path}"], content=["<encode>${content}</encode>"], space=["${id}"]);
-					//     CommitAsset(filePath=["${path}"], comment=["Hardcoded comment from the App Page editor"], space=["${id}"])
-					// `;
-				}
-
-				if (!pixel) {
-					throw new Error("Error missing pixel to get file");
-				}
-
-				const { errors } = await runPixel(pixel, insightId);
-
-				// bubble up the errors
-				for (const e of errors) {
-					throw new Error(e);
-				}
-
-				// reload the file
-				loadFile();
-				setIsModified(false);
-			} catch (e) {
-				notification.add({
-					color: "error",
-					message: e.message,
-				});
-
-				console.error(e);
-			} finally {
-				setIsLoading(false);
-			}
-		};
-
-		/**
-		 * Runs LLM pixel and manages LLM loading
-		 * @param string
-		 * @returns LLM response string
-		 */
-		const promptLLM = async (prompt: string) => {
-			// ideally add filetype to LLM pixel so it does not have to be in prompt string
-			// some formatting issues in return pixel including triple quotes and infrequent cutoffs in response string
-
-			try {
-				setIsLoading(true);
-
-				if (!agentModelEngine) {
-					throw new Error("No Agent Model Engine");
-				}
-
-				const response = await runPixel(
-					`LLM(engine = "${agentModelEngine}", command = "${prompt}", paramValues = [ {} ] );`,
-					insightId,
-				);
-
-				const LLMResponse = response.pixelReturn[0].output["response"];
-				let trimmedStarterCode = LLMResponse.replace(/^```|```$/g, ""); // trims off any triple quotes from backend
-
-				trimmedStarterCode = trimmedStarterCode.substring(
-					trimmedStarterCode.indexOf("\n") + 1,
-				);
-
-				return trimmedStarterCode;
-			} catch {
-				notification.add({
-					color: "error",
-					message: "Failed response from AI Code Generator",
-				});
-			} finally {
-				setIsLoading(false);
-			}
-		};
-
-		/**
-		 * Hanlder for setting content, isModified and onChange
-		 */
-		const handleEditorOnChange = (value) => {
-			setContent(value);
-			setIsModified(value !== initialContent);
-			onChange(value, isModified);
-		};
-
-		let syntaxSetupDone = false;
-
-		const setupSyntax = (monaco, language: string) => {
-			if (syntaxSetupDone) return;
-			syntaxSetupDone = true;
-
-			const config = languageConfigs[language];
-			if (!config) return;
-
-			// Setup Monarch tokens
-			if (config.monarchTokensProvider) {
-				monaco.languages.setMonarchTokensProvider(
-					language,
-					config.monarchTokensProvider,
-				);
-			}
-
-			// IntelliSense
-			if (typeof config.completionItems === "function") {
-				monaco.languages.registerCompletionItemProvider(language, {
-					triggerCharacters: [".", "("],
-					provideCompletionItems: (model, position) => {
-						const word = model.getWordUntilPosition(position);
-						const range = {
-							startLineNumber: position.lineNumber,
-							endLineNumber: position.lineNumber,
-							startColumn: word.startColumn,
-							endColumn: word.endColumn,
-						};
-
-						return {
-							suggestions: config
-								.completionItems(monaco)
-								.map((s) => ({
-									...s,
-									range,
-								})),
-						};
-					},
-				});
-			}
-
-			// Custom Theme
-			if (config.theme) {
-				monaco.editor.defineTheme(
-					`${language}-playground`,
-					config.theme,
-				);
-				monaco.editor.setTheme(`${language}-playground`);
-			}
-		};
-
-		/**
-		 * Handler called when the editor is mounted
-		 */
-		const onEditorMount: OnMount = (editor, monaco) => {
-			editorRef.current = editor;
-			if (IS_PRODUCTION) return;
-
-			setupSyntax(monaco, fileLanguage);
-			monaco.editor.setTheme(`${fileLanguage}-playground` || "vs");
-			// prevents redundant additions of new dropdown action
-			if (LLMActionAdded === false) {
-				setLLMActionAdded(true);
-				editor.addCommand(
-					monaco.KeyMod.CtrlCmd | monaco.KeyCode.KeyS,
-					() => {
-						saveFile();
-					},
-				); // use editor's api to use built in keyboard shortcuts to handle differnt OS's save
-				editor.addAction({
-					contextMenuGroupId: "1_modification",
-					contextMenuOrder: 1,
-					id: "prompt-LLM",
-					label: "Generate Code",
-					keybindings: [
-						monaco.KeyMod.CtrlCmd |
-							monaco.KeyMod.Shift |
-							monaco.KeyCode.KeyG,
-					],
-
-					run: async (editor) => {
-						const selection = editor.getSelection();
-						const selectedText = editor
-							.getModel()
-							.getValueInRange(selection);
-
-						const LLMReturnText = await promptLLM(
-							`Create code for a ${fileLanguage} file with the user prompt: ${selectedText}`, // filetype should be sent as param to LLM
-						);
-
-						editor.executeEdits("custom-action", [
-							{
-								range: new monaco.Range(
-									selection.endLineNumber + 2,
-									1,
-									selection.endLineNumber + 2,
-									1,
-								),
-								text: `\n\n${LLMReturnText}\n`,
-								forceMoveMarkers: true,
-							},
-						]);
-
-						editor.setSelection(
-							new monaco.Range(
-								selection.endLineNumber + 3,
-								1,
-								selection.endLineNumber +
-									2 +
-									LLMReturnText.split("\n").length,
-								1,
-							),
-						);
-					},
-				});
-				editor.addAction({
-					contextMenuGroupId: "1_modification",
-					contextMenuOrder: 2,
-					id: "toggle-word-wrap",
-					label: "Toggle Word Wrap",
-					keybindings: [monaco.KeyMod.Alt | monaco.KeyCode.KeyZ],
-					run: async (editor) => {
-						wordWrapRef.current = !wordWrapRef.current;
-						editor.updateOptions({
-							wordWrap: wordWrapRef.current ? "on" : "off",
-						});
-					},
-				});
-			}
-		};
-		const beforeMount = (monaco) => {
-			setupSyntax(monaco, fileLanguage);
-		};
-
-		return (
-			<StyledContainer>
-				<Suspense
-					fallback={
-						<LoadingScreen.Trigger description="Loading..." />
-					}
-				>
-					{isLoading ? (
-						<LoadingScreen.Trigger description="Loading..." />
-					) : (
-						<Editor
-							width={"100%"}
-							height={"100%"}
-							value={content}
-							language={fileLanguage}
-							options={{
-								readOnly: false,
-							}}
-							onChange={handleEditorOnChange}
-							onMount={onEditorMount}
-							beforeMount={beforeMount}
-						/>
-					)}
-				</Suspense>
-			</StyledContainer>
-		);
-	},
-);
