--- packages/client/src/components/common/File/FileEditorLanguageConfig.ts
@@ -1,301 +1,0 @@
-export const languageConfigs = {
-	java: {
-		monarchTokensProvider: {
-			defaultToken: "invalid",
-			keywords: [
-				"abstract",
-				"continue",
-				"for",
-				"new",
-				"switch",
-				"assert",
-				"default",
-				"goto",
-				"package",
-				"synchronized",
-				"boolean",
-				"do",
-				"if",
-				"private",
-				"this",
-				"break",
-				"double",
-				"implements",
-				"protected",
-				"throw",
-				"byte",
-				"else",
-				"import",
-				"public",
-				"throws",
-				"case",
-				"enum",
-				"instanceof",
-				"return",
-				"transient",
-				"catch",
-				"extends",
-				"int",
-				"short",
-				"try",
-				"char",
-				"final",
-				"interface",
-				"static",
-				"void",
-				"class",
-				"finally",
-				"long",
-				"strictfp",
-				"volatile",
-				"const",
-				"float",
-				"native",
-				"super",
-				"while",
-			],
-			typeKeywords: [
-				"byte",
-				"short",
-				"int",
-				"long",
-				"char",
-				"float",
-				"double",
-				"boolean",
-				"void",
-			],
-			operators: [
-				"=",
-				">",
-				"<",
-				"!",
-				"~",
-				"?",
-				":",
-				"==",
-				"<=",
-				">=",
-				"!=",
-				"&&",
-				"||",
-				"++",
-				"--",
-				"+",
-				"-",
-				"*",
-				"/",
-				"&",
-				"|",
-				"^",
-				"%",
-				"<<",
-				">>",
-				">>>",
-				"+=",
-				"-=",
-				"*=",
-				"/=",
-				"&=",
-				"|=",
-				"^=",
-				"%=",
-				"<<=",
-				">>=",
-				">>>=",
-			],
-			symbols: /[=><!~?:&|+\-*/^%]+/,
-			escapes: /\\(?:[btnfr"'\\]|u[0-9A-Fa-f]{4})/,
-			tokenizer: {
-				root: [
-					[/\b[a-zA-Z_][\w$]*(?=\s*\()/, "function"],
-					[
-						/[a-z_$][\w$]*/,
-						{
-							cases: {
-								"@keywords": "keyword",
-								"@default": "identifier",
-							},
-						},
-					],
-					[/[A-Z][\w$]*/, "type.identifier"],
-					{ include: "@whitespace" },
-					[/[{}()[\]]/, "@brackets"],
-					[
-						/(@symbols)/,
-						{
-							cases: {
-								"@operators": "operator",
-								"@default": "",
-							},
-						},
-					],
-					[/\d+\.\d+([eE][-+]?\d+)?[fFdD]?/, "number.float"],
-					[/0[xX][0-9a-fA-F]+[Ll]?/, "number.hex"],
-					[/\d+[lL]?/, "number"],
-					[/[;,.]/, "delimiter"],
-					[
-						/"/,
-						{
-							token: "string.quote",
-							bracket: "@open",
-							next: "@string",
-						},
-					],
-				],
-				comment: [
-					[/[^/*]+/, "comment"],
-					[/\*\//, "comment", "@pop"],
-					[/[/*]/, "comment"],
-				],
-				string: [
-					[/[^\\"]+/, "string"],
-					[/@escapes/, "string.escape"],
-					[/\\./, "string.escape.invalid"],
-					[
-						/"/,
-						{
-							token: "string.quote",
-							bracket: "@close",
-							next: "@pop",
-						},
-					],
-				],
-				whitespace: [
-					[/[ \t\r\n]+/, ""],
-					[/\/\*/, "comment", "@comment"],
-					[/\/\/.*$/, "comment"],
-				],
-			},
-		},
-		completionItems: (monaco) => [
-			{
-				label: "System.out.println",
-				kind: monaco.languages.CompletionItemKind.Function,
-				insertText: "System.out.println($1);",
-				insertTextRules:
-					monaco.languages.CompletionItemInsertTextRule
-						.InsertAsSnippet,
-				documentation: "Prints a message to the console.",
-			},
-			{
-				label: "public class",
-				kind: monaco.languages.CompletionItemKind.Snippet,
-				insertText: [
-					"public class ${1:ClassName} {",
-					"    public static void main(String[] args) {",
-					"        $0",
-					"    }",
-					"}",
-				].join("\n"),
-				insertTextRules:
-					monaco.languages.CompletionItemInsertTextRule
-						.InsertAsSnippet,
-				documentation: "Java class with main method.",
-			},
-		],
-		theme: {
-			base: "vs",
-			inherit: true,
-			rules: [
-				{ token: "keyword", foreground: "0000FF", fontStyle: "bold" },
-				{ token: "type.identifier", foreground: "2B91AF" },
-				{ token: "function", foreground: "B58B00" },
-				{ token: "comment", foreground: "008000", fontStyle: "italic" },
-				{ token: "string", foreground: "A31515" },
-				{ token: "number", foreground: "098658" },
-				{ token: "operator", foreground: "000000" },
-			],
-			colors: {},
-		},
-	},
-	python: {
-		completionItems: (monaco) => [
-			{
-				label: "print",
-				kind: monaco.languages.CompletionItemKind.Function,
-				insertText: 'print(${1:"Hello, world!"})',
-				insertTextRules:
-					monaco.languages.CompletionItemInsertTextRule
-						.InsertAsSnippet,
-				documentation: "Print to the console.",
-			},
-			{
-				label: "for loop",
-				kind: monaco.languages.CompletionItemKind.Snippet,
-				insertText: [
-					"for ${1:item} in ${2:iterable}:",
-					"    ${0:# do something}",
-				].join("\n"),
-				insertTextRules:
-					monaco.languages.CompletionItemInsertTextRule
-						.InsertAsSnippet,
-				documentation: "For loop in Python.",
-			},
-			{
-				label: "if-else",
-				kind: monaco.languages.CompletionItemKind.Snippet,
-				insertText: [
-					"if ${1:condition}:",
-					"    ${0:# do something}",
-					"else:",
-					"    # handle else",
-				].join("\n"),
-				insertTextRules:
-					monaco.languages.CompletionItemInsertTextRule
-						.InsertAsSnippet,
-				documentation: "If-else block in Python.",
-			},
-			{
-				label: "def function",
-				kind: monaco.languages.CompletionItemKind.Snippet,
-				insertText: [
-					"def ${1:function_name}(${2:args}):",
-					"    ${0:pass}",
-				].join("\n"),
-				insertTextRules:
-					monaco.languages.CompletionItemInsertTextRule
-						.InsertAsSnippet,
-				documentation: "Define a Python function.",
-			},
-			{
-				label: "class",
-				kind: monaco.languages.CompletionItemKind.Snippet,
-				insertText: [
-					"class ${1:ClassName}:",
-					"    def __init__(self, ${2:args}):",
-					"        ${0:pass}",
-				].join("\n"),
-				insertTextRules:
-					monaco.languages.CompletionItemInsertTextRule
-						.InsertAsSnippet,
-				documentation: "Define a class with a constructor.",
-			},
-			{
-				label: "while loop",
-				kind: monaco.languages.CompletionItemKind.Snippet,
-				insertText: [
-					"while ${1:condition}:",
-					"    ${0:# do something}",
-				].join("\n"),
-				insertTextRules:
-					monaco.languages.CompletionItemInsertTextRule
-						.InsertAsSnippet,
-				documentation: "While loop in Python.",
-			},
-		],
-		theme: {
-			base: "vs",
-			inherit: true,
-			rules: [
-				{ token: "identifier", foreground: "B58B00" },
-				{ token: "keyword", foreground: "0000FF", fontStyle: "bold" },
-				{ token: "comment", foreground: "008000", fontStyle: "italic" },
-				{ token: "string", foreground: "A31515" },
-				{ token: "number", foreground: "098658" },
-				{ token: "operator", foreground: "000000" },
-			],
-			colors: {},
-		},
-	},
-};
