--- packages/client/src/components/common/File/index.ts
@@ -1,5 +1,0 @@
-export * from "./AddFileOverlay";
-export * from "./CreateFileOverlay";
-export * from "./DeleteFileOverlay";
-export * from "./FileEditor";
-export * from "./FileExplorer";
