+++ packages/client/src/components/common/index.ts
@@ -1,4 +1,3 @@
 export * from "./ErrorBoundary";
-export * from "./File";
 export * from "./MarkdownEditor";
 export * from "./TextEditor";
