+++ packages/client/src/components/workspace/index.ts
@@ -1,3 +1,4 @@
 export * from "./PreviewOverlay";
 export * from "./panels";
-export * from "./Workspace";
+export * from "./WorkspaceManager";
+export * from "./workspace"
