--- packages/client/src/components/workspace/panels/FileEditorPanel.tsx
@@ -1,240 +1,0 @@
-import {
-	ContentCopyOutlined,
-	RefreshOutlined,
-	SaveOutlined,
-} from "@mui/icons-material";
-import { observer } from "mobx-react-lite";
-import { useRef, useState } from "react";
-import {
-	Button,
-	IconButton,
-	Modal,
-	Stack,
-	Typography,
-	useNotification,
-} from "@semoss/ui";
-import { FileEditor, type FileEditorRefDef } from "@/components/common";
-import { FlexLayout } from "@/components/flex-layout";
-import { useWorkspace } from "@/hooks";
-import { Panel } from "./Panel";
-
-interface FileEditorPanelProps {
-	/** Path to the file location */
-	path: string;
-}
-
-export const FileEditorPanel = observer((props: FileEditorPanelProps) => {
-	const { path } = props;
-	const { workspace } = useWorkspace();
-	const notification = useNotification();
-
-	const [isModified, setIsModified] = useState(false);
-
-	const fileEditorRef = useRef<FileEditorRefDef>(null);
-	const [refreshContent, setRefreshContent] = useState(1);
-	const [saveBeforeRefresh, setSaveBeforeRefresh] = useState(false);
-
-	// get the name
-	const name = path.split("/").pop();
-
-	/**
-	 * Triggered when a file is changed
-	 * @param isModified - isModified
-	 */
-	const onFileEditorChange = (isModified: boolean) => {
-		try {
-			// update
-			setIsModified(isModified);
-
-			// update the tabs
-			updatePanels(isModified);
-		} catch (e) {
-			notification.add({
-				color: "error",
-				message: e,
-			});
-		}
-	};
-
-	/**
-	 * Triggered when a file is changed
-	 * @param isModified - isModified
-	 */
-	const updatePanels = (isModified: boolean) => {
-		try {
-			// get the model
-			const model = workspace.model;
-			if (!model) {
-				throw new Error("Missing model");
-			}
-
-			// visit the notes, and see if it exists
-			model.visitNodes((node) => {
-				// check if it is a tabNode
-				if (node instanceof FlexLayout.TabNode) {
-					// it needs to be a file-editor
-					const component = node.getComponent();
-					if (component !== "file-editor") {
-						return;
-					}
-
-					// path and space need to match
-					const config = node.getConfig();
-					if (path !== config.path) {
-						return;
-					}
-
-					const id = node.getId();
-					if (isModified) {
-						model.doAction(
-							FlexLayout.Actions.renameTab(id, `${name}*`),
-						);
-					} else {
-						model.doAction(
-							FlexLayout.Actions.renameTab(id, `${name}`),
-						);
-					}
-				}
-			});
-		} catch (e) {
-			notification.add({
-				color: "error",
-				message: e,
-			});
-		}
-	};
-
-	/**
-	 * Copy the path to the clipboard
-	 */
-	const copyPath = async () => {
-		try {
-			await navigator.clipboard.writeText(path);
-
-			notification.add({
-				color: "success",
-				message: "Successfully copied path",
-			});
-		} catch {
-			notification.add({
-				color: "error",
-				message: "Unable to copy path",
-			});
-		}
-	};
-
-	return (
-		<Panel
-			actions={
-				<>
-					<IconButton
-						size={"small"}
-						color={"default"}
-						title={`Copy path - ${path}`}
-						onClick={(e) => {
-							e.preventDefault();
-							e.stopPropagation();
-							copyPath();
-						}}
-					>
-						<ContentCopyOutlined fontSize="inherit" />
-					</IconButton>
-					<IconButton
-						size={"small"}
-						color={"default"}
-						title={`Refresh File - ${path}`}
-						onClick={(e) => {
-							e.preventDefault();
-							e.stopPropagation();
-							if (isModified) {
-								setSaveBeforeRefresh(true);
-							} else {
-								setRefreshContent((prev) => prev + 1);
-							}
-							// copyPath();
-						}}
-					>
-						<RefreshOutlined fontSize="inherit" />
-					</IconButton>
-					<Stack flex={1}>&nbsp;</Stack>
-					<IconButton
-						size={"small"}
-						color={"default"}
-						title={"Save"}
-						disabled={!isModified}
-						onClick={(e) => {
-							e.preventDefault();
-							e.stopPropagation();
-
-							// trigger the save
-							fileEditorRef.current?.saveFile();
-						}}
-					>
-						<SaveOutlined fontSize="inherit" />
-					</IconButton>
-				</>
-			}
-		>
-			<FileEditor
-				ref={fileEditorRef}
-				type={"app"}
-				space={workspace.appId}
-				insightId={workspace.insightId}
-				path={path}
-				agentModelEngine={workspace.agentModelEngine}
-				refreshContent={refreshContent}
-				onChange={(_content, isModified) => {
-					onFileEditorChange(isModified);
-				}}
-			/>
-			<Modal
-				open={saveBeforeRefresh}
-				maxWidth="sm"
-				fullWidth
-				onClose={() => setSaveBeforeRefresh(false)}
-				aria-labelledby="export-modal-title"
-				aria-describedby="export-modal-description"
-			>
-				<Modal.Title>
-					<Typography id={"export-modal-title"} variant="h6">
-						File Content Modified
-					</Typography>
-				</Modal.Title>
-				<Modal.Content>
-					<Typography
-						id={"export-modal-description"}
-						variant="body1"
-						sx={{ mb: 2 }}
-					>
-						File content is modified. Do you want to reload the
-						file?
-					</Typography>
-				</Modal.Content>
-				<Modal.Actions>
-					<Button
-						variant="contained"
-						color="primary"
-						onClick={() => {
-							setRefreshContent((prev) => prev + 1);
-							setSaveBeforeRefresh(false);
-							setIsModified(false);
-							// update the tabs
-							updatePanels(false);
-						}}
-					>
-						Yes
-					</Button>
-					<Button
-						variant="outlined"
-						color="secondary"
-						onClick={() => {
-							setSaveBeforeRefresh(false);
-						}}
-					>
-						No
-					</Button>
-				</Modal.Actions>
-			</Modal>
-		</Panel>
-	);
-});
