+++ packages/client/src/components/workspace/panels/index.tsx
@@ -1,6 +1,3 @@
-export * from "./EmptyFilePanel";
-export * from "./FileEditorPanel";
-export * from "./FileExplorerPanel";
 export * from "./Panel";
 export * from "./SettingsPanel";
 export * from "./TerminalPanel";
