+++ packages/client/src/constants.ts
@@ -47,10 +47,8 @@ export const PERMISSION_DESCRIPTION_MAP: Record<
 };
 
 export const MCP = {
-	DRIVER_PATHS: ["/py/mcp_driver.py", "py/mcp_driver.py"],
+	DRIVER_PATHS: ["/py/mcp_driver.py"],
 	JSON_PATHS: [
-		"mcp/py_mcp.json",
-		"mcp/pixel_mcp.json",
 		"/mcp/py_mcp.json",
 		"/mcp/pixel_mcp.json",
 	],
